The overlays stop hand-declaring the resident functions they call (#2495)
Seven overlay sources declared resident functions themselves, most of them
under address names for functions the resident tree has long since named:
func_80035BE4 for TextBox_Create, func_80035B7C for TextBox_Destroy,
func_8003F87C for SaveData_RequestWrite, and so on. Twenty-six of those
declarations are gone. The calls now use the resident names through the
resident headers, and the four overlay linker-symbol files gain the
addresses the renamed calls need. The files are set_location.c, shop.c,
name_entry_runtime.c, name_entry_main.c, screen_runtime.c, frontend.c and
trade_update.c.

Five resident sources had no header to consume, so they get one:
func_8005922C.h, model_slot_data.h, func_80061008.h, func_800610E0.h and
func_800611D0.h. func_8005922C's declaration moves out of model.h into its
own header, which its definition now includes; model.h was the only place it
was declared, and its definition never saw it. The two resident callers
include the new header.

func_80024DC8 now takes s32 parameters rather than (s8, s8, u16, u16).
The Free Duel overlay passes a computed grid index as the second argument,
and under the s8 prototype FreeDuel_UpdateScreen grows an sll/sra pair
retail does not have (206 -> 208 instructions). The definition stores
through sb/sh whatever the parameter type, and both resident callers pass
constants, so the resident image does not move either.

func_800291E0.h no longer claims that overlay sources cannot consume
resident headers. They can and do; shop.c now includes this one and casts
the result to its own PasswordCardPreviewView view.

Left alone:
- func_800857C0, the SetGeomScreen wrapper notes/psyq.md deliberately
  keeps address-named.
- The six callees in alternate_pick_exit.c and alternate_location_tick.c,
  whose addresses are not in the resident inventory.

Resident image and five overlay hashes unchanged.

+++ src/game/func_80024DC8.c
@@ -9,7 +9,7 @@ extern u16 D_8009B372;
 extern u16 D_8009B374;
 extern u8 D_8009B26C;
 
-void func_80024DC8(s8 arg0, s8 arg1, u16 arg2, u16 arg3)
+void func_80024DC8(s32 arg0, s32 arg1, s32 arg2, s32 arg3)
 {
     D_8009B36A = 0x7270;
     D_8009B374 = 0x7280;
